# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Books Database App API",
      "description": "All backend APIs",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:8000",
    "basePath": "/api/",
...
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "securityDefinitions": {
      "Bearer":{
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "IMPORTANT!!! Please use this format: Bearer: Token"
      }
  },
    "paths": {
...
      "/books": {
        "get": {
          "summary": "Get all books",
          "tags": [
            "Books"
          ],
          "description": "Returns all books in DB",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Books returned",
              "schema": {
                "$ref": "#/definitions/Book"
              }
            }
          }
        },
        "post": {
          "security": [{"Bearer": []}] ,
          "summary": "Create a books",
          "tags": [
            "Books"
          ],
          "description": "Creates a book in database",
          "parameters": [
            {
              "name": "book",
              "in": "body",
              "description": "get books",
              "schema": {
                "$ref":"#/definitions/Book"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "Created a new book",
              "schema": {
                "$ref": "#/definitions/Book"
              }
            },
            "401":{
              "description": "Unauthorised!"
            }
          }
        }
      },
...
      "/books/{id}/authors": {
        "get": {
          "summary": "Get a list of authors for {id} book",
          "tags": [
            "Books"
          ],
          "description": "Get a list of authors for {id} book",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Book id",
              "schema": {
                "type": "string"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "List of authoirs for book {id}",
              "schema": {
                "$ref": "#/definitions/Author"
              }
            },
            "404":{
              "description": "Book not found!"
            }
          }
        },
        "post": {
          "security": [{"Bearer": []}] ,
          "summary": "Add new author to {id} book",
          "tags": [
            "Books"
          ],
          "description": "Add a new author to book {id}",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Book id",
              "schema": {
                "type":"string"
              }
            },
            {
              "in": "body",
              "name": "author",
              "description": "Author id",
              "schema": {
                  "properties": {
                    "author" : {
                      "type": "array",
                "items":{"type":"string"}
                    }
                  }
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Added a new author to book {id}",
              "schema": {
                "$ref": "#/definitions/Book"
              }
            },
            "401":{
              "description": "Unauthorised!"
            },
            "404":{
              "description": "Book not found!"
            }
          }
        }
      },
...
      "/authors": {
        "get": {
          "summary": "Get list of authors",
          "tags": [
            "Authors"
          ],
          "description": "Returns all authors in DB",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Authors returned",
              "schema": {
                "$ref": "#/definitions/Author"
              }
            }
          }
        },
        "post": {
          "security": [{"Bearer": []}] ,
          "summary": "Create new author",
          "tags": [
            "Authors"
          ],
          "description": "Creates an author in database",
          "parameters": [
            {
              "in": "body",
              "name": "author",
              "description": "Post author",
              "schema": {
                "$ref": "#/definitions/Author"
              }
            }

          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Created a new author",
              "schema": {
                "$ref": "#/definitions/Author"
              }
            },
            "401":{
              "description": "Unauthorised!"
            }
          }
        }
      },
...
      "/authors/{idAuthor}/books": {
        "get": {
          "summary": "Get a list of books for {idAuthor} author",
          "tags": [
            "Authors"
          ],
          "description": "Get a list of books for {idAuthor} author",
          "parameters": [
            {
              "in": "path",
              "name": "idAuthor",
              "description": "Author id",
              "schema": {
                "type": "string"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "List of books for {idAuthor} author",
              "schema": {
                "$ref": "#/definitions/Book"
              }
            },
            "404":{
              "description": "Author not found!"
            }
          }
        },
        "post": {
          "security": [{"Bearer": []}] ,
          "summary": "Add new book to {idAuthor} Author",
          "tags": [
            "Authors"
          ],
          "description": "Add a new book to {idAuthor} Author",
          "parameters": [
            {
              "in": "path",
              "name": "idAuthor",
              "description": "Author id",
              "schema": {
                "type": "string"
              }
            },
              {
                "in": "body",
                "name": "books",
                "description": "Book id",
                "schema": {
                  "properties": {
                    "books" : {
                      "type": "array",
                      "items": {"type":"string"}
                    }
                  }
                }
              }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Added a new book to {idAuthor} author",
              "schema": {
                "$ref": "#/definitions/Author"
              }
            },
            "401":{
              "description": "Unauthorised!"
            },
            "404":{
              "description": "Book not found!"
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Book": {
        "properties": {
          "isbn" : {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "pages": {
            "type": "number"
          },
          "published": {
            "type": "number"
          },
          "authors": {
            "type": "array",
            "items":{"type": "string"}
              
          },
          "image": {
            "type": "string"
          }
        }
      },
      "Author": {
        "properties": {
          "id" : {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "books": {
            "type": "array",
            "items":{"type": "string"}
              
          },
          "image": {
            "type": "string"
          }
        }
      }
    }
  }